+++ docs/.vitepress/theme/data/config.ts
@@ -1,0 +1,5 @@
+import { SiteConfig } from "vitepress";
+
+const config: SiteConfig = globalThis.VITEPRESS_CONFIG;
+
+export default config;
